refactor(story): inline WithProvider.tsx into preview.tsx
